Reorganize and update dataloader call
