+++ utils/wandb_logging/log_dataset.py
@@ -1,54 +1,66 @@
 import argparse
+import logging
 import os
+import random
+import time
+from pathlib import Path
+from threading import Thread
+from warnings import warn
 
+import numpy as np
 import torch
 import yaml
+import wandb
+from wandb_utils import WandbLogger
 
-from utils.datasets import create_dataloader
 from utils.general import check_dataset, colorstr
 from utils.torch_utils import torch_distributed_zero_first
-from wandb_utils import WandbLogger
+from utils.datasets import create_dataloader
 
 WANDB_ARTIFACT_PREFIX = 'wandb-artifact://'
 
-
 def create_dataset_artifact(opt):
     with open(opt.data) as f:
-        data_dict = yaml.load(f, Loader=yaml.SafeLoader)  # data dict
+        data_dict = yaml.load(f, Loader=yaml.FullLoader)  # data dict
     wandb_logger = WandbLogger(opt, '', None, data_dict, job_type='create_dataset')
     # Hyperparameters
     with open(opt.hyp) as f:
-        hyp = yaml.load(f, Loader=yaml.SafeLoader)  # load hyps
+        hyp = yaml.load(f, Loader=yaml.FullLoader)  # load hyps
+
     with torch_distributed_zero_first(-1):
         check_dataset(data_dict)  # check
-
+    train_path = data_dict['train']
+    test_path = data_dict['val']
     nc, names = (1, ['item']) if opt.single_cls else (int(data_dict['nc']), data_dict['names'])
     imgsz, batch_size = opt.img_size, opt.batch_size
-    trainloader, trainset = create_dataloader(data_dict['train'], imgsz, batch_size, stride=32, opt=opt,
-                                              hyp=hyp, cache=False, rect=opt.rect, rank=-1,
-                                              world_size=opt.world_size, workers=opt.workers,
-                                              image_weights=opt.image_weights, quad=opt.quad,
-                                              prefix=colorstr('train: '))
-
-    testloader = create_dataloader(data_dict['val'], imgsz, batch_size, stride=32, opt=opt,  # testloader
-                                   hyp=hyp, cache=False, rect=True, rank=-1,
-                                   world_size=opt.world_size, workers=opt.workers,
-                                   pad=0.5, prefix=colorstr('val: '))[0]
+    assert len(names) == nc, '%g names found for nc=%g dataset in %s' % (len(names), nc, opt.data)  # check
+    trainloader, trainset = create_dataloader(train_path, imgsz, batch_size, stride=32, opt=opt,
+                                            hyp=hyp, cache=False, rect=opt.rect, rank=-1,
+                                            world_size=opt.world_size, workers=opt.workers,
+                                            image_weights=opt.image_weights, quad=opt.quad, prefix=colorstr('train: '))
+    mlc = np.concatenate(trainset.labels, 0)[:, 0].max()  # max label class
+    assert mlc < nc, 'Label class %g exceeds nc=%g in %s. Possible class labels are 0-%g' % (mlc, nc, opt.data, nc - 1)
+    
+    testloader = create_dataloader(test_path, imgsz, batch_size, stride=32, opt=opt,  # testloader
+                                       hyp=hyp, cache=False, rect=True, rank=-1,
+                                       world_size=opt.world_size, workers=opt.workers,
+                                       pad=0.5, prefix=colorstr('val: '))[0]
     names_to_ids = {k: v for k, v in enumerate(names)}
 
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
     wandb_logger.log_dataset_artifact(trainloader, device, names_to_ids, name='train')
     wandb_logger.log_dataset_artifact(testloader, device, names_to_ids, name='val')
-    # Update/Create new config file with links to artifact
+    #Update/Create new config file with links to artifact
     data_dict['train'] = WANDB_ARTIFACT_PREFIX + opt.project + '/train'
     data_dict['val'] = WANDB_ARTIFACT_PREFIX + opt.project + '/val'
-    ouput_data_config = opt.data if opt.overwrite_config else opt.data.replace('.', '_wandb.')
-    data_dict.pop('download', None)  # Don't download the original dataset. Use artifacts
+    ouput_data_config = opt.data if opt.overwrite_config else opt.data.replace('.','_wandb.')
+    data_dict.pop('download',None) #Don't download the original dataset. Use artifacts
     with open(ouput_data_config, 'w') as fp:
         yaml.dump(data_dict, fp)
     print("New Config file => ", ouput_data_config)
-
-
+    
+    
+    
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--data', type=str, default='data/coco128.yaml', help='data.yaml path')
